feat: Add UpliftPage, ChecklistPage, and PathwayPage components
feat: Add Compound Calculator page to Finances section

feat: Add OperationsHistoryChart component to TradingJournalPage

refactor: Adjust height of OperationsHistory component

refactor: Update Calendar component to handle no trades

feat: Add ConfirmDialog component to common directory

feat: Add hint for adding trades in EditPanel component
