+++ src/components/copyright/Copyright.types.ts
