+++ src/components/copyright/index.ts
