+++ src/pages/SignInPage.tsx
@@ -5,7 +5,7 @@ import {
   browserSessionPersistence,
 } from 'firebase/auth';
 import { auth } from '../firebase';
-import { AutoLogRoute } from 'src/components/common';
+import { AutoLogRoute } from 'src/components';
 import SignIn, { SignInFormInputs } from 'src/components/pages/sign-in';
 import { useNavigate } from 'react-router-dom';
 import { handleAuthError } from 'src/utils/auth';
