+++ src/pages/client/_finances/PersonalFinances/components/FixedExpenseEditDialog.tsx
@@ -5,18 +5,19 @@ import {
   DialogContent,
   DialogActions,
   Button,
+  TextField,
   IconButton,
   useTheme,
   useMediaQuery,
   Grid,
   Box,
+  MenuItem,
   SxProps,
   Tooltip,
   Accordion,
   AccordionDetails,
   AccordionSummary,
   Typography,
-  TextField,
 } from '@mui/material';
 import {
   Edit as EditIcon,
@@ -26,55 +27,48 @@ import {
 import { useForm, Controller, useFieldArray } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
-import { IDebt } from 'src/models/finances';
+import { CurrencyField } from 'src/components/forms';
+import { IFixedExpense } from 'src/models/finances';
 import { useTranslation } from 'react-i18next';
-import { CurrencyField, PercentageField } from 'src/components/common/forms';
-import { DatePicker } from '@mui/x-date-pickers';
-import dayjs from 'dayjs';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import { formatCurrency, formatPercentage } from 'src/utils';
+import dayjs from 'dayjs';
+import { formatCurrency } from 'src/utils';
+import { DatePicker } from '@mui/x-date-pickers';
 
 interface Props {
-  onSubmit: (data: IDebt[]) => void;
-  data: IDebt[];
+  onSubmit: (data: IFixedExpense[]) => void;
+  data: IFixedExpense[];
   sx?: SxProps;
   loading?: boolean;
 }
 
-const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
+const FixedExpenseEditDialog = ({ onSubmit, data, loading, sx }: Props) => {
   const { t } = useTranslation();
   const schema = useMemo(
     () =>
       yup.object().shape({
-        debts: yup.array().of(
+        expenses: yup.array().of(
           yup.object().shape({
             name: yup
               .string()
               .nonNullable()
               .required(t('commonValidations.required'))
               .max(64),
-            pendingDebt: yup
+            amount: yup
               .number()
               .nonNullable()
               .required(t('commonValidations.required'))
               .typeError(t('commonValidations.required'))
               .moreThan(-1),
-            minimumPayment: yup
-              .number()
+            expenseType: yup
+              .string()
               .nonNullable()
-              .required(t('commonValidations.required'))
-              .typeError(t('commonValidations.required'))
-              .moreThan(-1),
-            annualInterest: yup
-              .number()
-              .nonNullable()
-              .required(t('commonValidations.required'))
-              .typeError(t('commonValidations.required'))
-              .moreThan(-1),
-            startDate: yup
-              .date()
-              .nonNullable()
+              .oneOf(
+                ['primary', 'secondary', 'single'],
+                t('commonValidations.type')
+              )
               .required(t('commonValidations.required')),
+            singleDate: yup.date().nullable(),
           })
         ),
       }),
@@ -95,47 +89,47 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
   } = useForm({
     resolver: yupResolver(schema),
     defaultValues: {
-      debts: data,
+      expenses: data,
     },
   });
 
   const { fields, append, remove } = useFieldArray({
     control,
-    name: 'debts',
+    name: 'expenses',
   });
 
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
 
-  function _handleSubmit(data: { debts?: IDebt[] }) {
-    if (!data.debts) return;
+  function _handleSubmit(data: {
+    expenses?: Omit<IFixedExpense, 'startDate'>[];
+  }) {
+    if (!data.expenses) return;
 
     handleClose();
-    onSubmit(data.debts);
+    onSubmit(data.expenses.map((item) => ({ ...item, startDate: new Date() })));
   }
 
   useEffect(() => {
-    setValue('debts', data.sort((x) => x.pendingDebt).reverse());
+    setValue('expenses', data);
   }, [data, setValue]);
 
-  const debts = watch('debts');
+  const expenses = watch('expenses');
 
-  const handleOnNewDebt = () => {
+  function handleOnNewExpense() {
     append({
-      name: `${t('finances.personalFinances.header.debts.dialog.debt')} ${
-        fields.length + 1
-      }`,
-      pendingDebt: 0,
-      minimumPayment: 0,
-      annualInterest: 0,
-      startDate: dayjs() as any,
+      amount: 0,
+      expenseType: 'primary',
+      name: `${t('finances.personalFinances.header.fixedExpenses.dialog.fixedExpense')} ${fields.length + 1}`,
     });
     setExpanded(`panel${fields.length}`);
-  };
+  }
 
   return (
     <>
-      <Tooltip title={t('finances.personalFinances.header.debts.dialog.title')}>
+      <Tooltip
+        title={t('finances.personalFinances.header.fixedExpenses.dialog.title')}
+      >
         <Box sx={{ display: 'inline-block', ...sx }}>
           <IconButton onClick={handleOpen} disabled={loading}>
             <EditIcon fontSize="small" />
@@ -160,13 +154,15 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
             justifyContent: 'space-between',
           }}
         >
-          {t('finances.personalFinances.header.debts.dialog.title')}
+          {t('finances.personalFinances.header.fixedExpenses.dialog.title')}
           <Button
             type="button"
-            onClick={handleOnNewDebt}
+            onClick={handleOnNewExpense}
             startIcon={<AddIcon />}
           >
-            {t('finances.personalFinances.header.debts.dialog.addDebt')}
+            {t(
+              'finances.personalFinances.header.fixedExpenses.dialog.addFixedExpense'
+            )}
           </Button>
         </DialogTitle>
         <DialogContent>
@@ -190,36 +186,47 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
                       <Box sx={{ width: '75%' }} mr={1}>
                         <Typography
                           color={
-                            Object.values(errors?.debts?.[index] ?? {})
+                            Object.values(errors?.expenses?.[index] ?? {})
                               .length && expanded !== `panel${index}`
                               ? 'error.main'
                               : 'inherit'
                           }
                         >
-                          {debts![index].name}
+                          {expenses![index].name}
                         </Typography>
                         <Typography
                           variant="caption"
                           color={
-                            Object.values(errors?.debts?.[index] ?? {})
+                            Object.values(errors?.expenses?.[index] ?? {})
                               .length && expanded !== `panel${index}`
                               ? 'error.main'
                               : 'text.secondary'
                           }
                         >
-                          {formatPercentage(debts![index].annualInterest)}
+                          {
+                            <Box
+                              component="span"
+                              sx={{ textTransform: 'capitalize' }}
+                            >
+                              {expenses![index].expenseType !== 'single'
+                                ? expenses![index].expenseType
+                                : dayjs(expenses![index].singleDate).format(
+                                    'DD MMM YYYY'
+                                  )}
+                            </Box>
+                          }
                         </Typography>
                       </Box>
                       <Box display="flex" alignItems="center">
                         <Typography
                           color={
-                            Object.values(errors?.debts?.[index] ?? {})
+                            Object.values(errors?.expenses?.[index] ?? {})
                               .length && expanded !== `panel${index}`
                               ? 'error.main'
                               : 'text.secondary'
                           }
                         >
-                          {formatCurrency(debts![index].pendingDebt)}
+                          {formatCurrency(expenses![index].amount)}
                         </Typography>
                       </Box>
                     </>
@@ -230,76 +237,49 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
                   )}
                 </AccordionSummary>
                 <AccordionDetails>
-                  <Grid container spacing={2} alignItems="center">
+                  <Grid container spacing={2} alignItems="center" key={item.id}>
                     <Grid item xs={8}>
                       <Controller
-                        name={`debts.${index}.name`}
+                        name={`expenses.${index}.name`}
                         control={control}
                         render={({ field }) => (
                           <TextField
                             {...field}
-                            size="small"
                             label={t(
-                              'finances.personalFinances.header.debts.dialog.name'
+                              'finances.personalFinances.header.fixedExpenses.dialog.name'
                             )}
                             fullWidth
-                            error={!!errors?.debts?.[index]?.name}
+                            error={!!errors?.expenses?.[index]?.expenseType}
                             helperText={
-                              errors?.debts?.[index]?.name?.message || ' '
+                              errors?.expenses?.[index]?.expenseType?.message ||
+                              ' '
                             }
                             margin="dense"
-                            variant="filled"
+                            size="small"
                             inputProps={{ maxLength: 64 }}
+                            variant="filled"
                           />
                         )}
                       />
                     </Grid>
                     <Grid item xs={4}>
                       <Controller
-                        name={`debts.${index}.startDate`}
-                        control={control}
-                        render={({ field }) => (
-                          <DatePicker
-                            {...field}
-                            value={field.value ?? null}
-                            slotProps={{
-                              textField: {
-                                fullWidth: true,
-                                size: 'small',
-                                error: !!errors?.debts?.[index]?.startDate,
-                                helperText:
-                                  errors?.debts?.[index]?.startDate?.message ||
-                                  ' ',
-                                margin: 'dense',
-                                variant: 'filled',
-                              },
-                            }}
-                            label={t(
-                              'finances.personalFinances.header.debts.dialog.startDate'
-                            )}
-                            views={['year', 'month', 'day']}
-                          />
-                        )}
-                      />
-                    </Grid>
-                    <Grid item xs={4}>
-                      <Controller
-                        name={`debts.${index}.pendingDebt`}
+                        name={`expenses.${index}.amount`}
                         control={control}
                         render={({ field }) => (
                           <CurrencyField
                             {...field}
-                            size="small"
+                            value={field.value.toString()}
                             label={t(
-                              'finances.personalFinances.header.debts.dialog.pendingDebt'
+                              'finances.personalFinances.header.fixedExpenses.dialog.amount'
                             )}
                             fullWidth
-                            error={!!errors?.debts?.[index]?.pendingDebt}
+                            error={!!errors?.expenses?.[index]?.amount}
                             helperText={
-                              errors?.debts?.[index]?.pendingDebt?.message ||
-                              ' '
+                              errors?.expenses?.[index]?.amount?.message || ' '
                             }
                             margin="dense"
+                            size="small"
                             inputProps={{ min: 0 }}
                             variant="filled"
                           />
@@ -308,52 +288,75 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
                     </Grid>
                     <Grid item xs={4}>
                       <Controller
-                        name={`debts.${index}.minimumPayment`}
+                        name={`expenses.${index}.expenseType`}
                         control={control}
                         render={({ field }) => (
-                          <CurrencyField
+                          <TextField
                             {...field}
-                            size="small"
                             label={t(
-                              'finances.personalFinances.header.debts.dialog.minPayment'
+                              'finances.personalFinances.header.fixedExpenses.dialog.expenseType'
                             )}
                             fullWidth
-                            error={!!errors?.debts?.[index]?.minimumPayment}
+                            select
+                            error={!!errors?.expenses?.[index]?.expenseType}
                             helperText={
-                              errors?.debts?.[index]?.minimumPayment?.message ||
+                              errors?.expenses?.[index]?.expenseType?.message ||
                               ' '
                             }
                             margin="dense"
-                            inputProps={{ min: 0 }}
-                            variant="filled"
-                          />
-                        )}
-                      />
-                    </Grid>
-                    <Grid item xs={4}>
-                      <Controller
-                        name={`debts.${index}.annualInterest`}
-                        control={control}
-                        render={({ field }) => (
-                          <PercentageField
-                            {...field}
                             size="small"
-                            label={t(
-                              'finances.personalFinances.header.debts.dialog.anualInterest'
-                            )}
-                            fullWidth
-                            error={!!errors?.debts?.[index]?.annualInterest}
-                            helperText={
-                              errors?.debts?.[index]?.annualInterest?.message ||
-                              ' '
-                            }
-                            margin="dense"
-                            inputProps={{ min: 0 }}
                             variant="filled"
-                          />
+                          >
+                            <MenuItem value="single">
+                              {t(
+                                'finances.personalFinances.header.fixedExpenses.dialog.expenseTypes.single'
+                              )}
+                            </MenuItem>
+                            <MenuItem value="primary">
+                              {t(
+                                'finances.personalFinances.header.fixedExpenses.dialog.expenseTypes.primary'
+                              )}
+                            </MenuItem>
+                            <MenuItem value="secondary">
+                              {t(
+                                'finances.personalFinances.header.fixedExpenses.dialog.expenseTypes.secondary'
+                              )}
+                            </MenuItem>
+                          </TextField>
                         )}
                       />
                     </Grid>
+                    {Boolean(expenses![index].expenseType === 'single') && (
+                      <Grid item xs={4}>
+                        <Controller
+                          name={`expenses.${index}.singleDate`}
+                          control={control}
+                          render={({ field }) => (
+                            <DatePicker
+                              {...field}
+                              value={field.value ?? null}
+                              slotProps={{
+                                textField: {
+                                  fullWidth: true,
+                                  size: 'small',
+                                  error:
+                                    !!errors?.expenses?.[index]?.singleDate,
+                                  helperText:
+                                    errors?.expenses?.[index]?.singleDate
+                                      ?.message || ' ',
+                                  margin: 'dense',
+                                  variant: 'filled',
+                                },
+                              }}
+                              label={t(
+                                'finances.personalFinances.header.fixedExpenses.dialog.date'
+                              )}
+                              views={['year', 'month', 'day']}
+                            />
+                          )}
+                        />
+                      </Grid>
+                    )}
                   </Grid>
                 </AccordionDetails>
               </Accordion>
@@ -362,10 +365,10 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
         </DialogContent>
         <DialogActions>
           <Button type="button" onClick={handleClose}>
-            {t('finances.personalFinances.header.debts.dialog.cancel')}
+            {t('finances.personalFinances.header.fixedExpenses.dialog.cancel')}
           </Button>
           <Button type="submit">
-            {t('finances.personalFinances.header.debts.dialog.save')}
+            {t('finances.personalFinances.header.fixedExpenses.dialog.save')}
           </Button>
         </DialogActions>
       </Dialog>
@@ -373,4 +376,4 @@ const DebtEditDialog = ({ onSubmit, data, sx, loading }: Props) => {
   );
 };
 
-export default DebtEditDialog;
+export default FixedExpenseEditDialog;
