+++ src/pages/client/_finances/PersonalFinances/index.ts
@@ -1,0 +1,1 @@
+export { default as PersonalFinancesPage } from './Page';
