+++ src/pages/client/_finances/Portfolio/index.ts
@@ -1,0 +1,1 @@
+export { default as PortfolioPage } from './Page';
