+++ src/pages/client/_finances/TradingJournal/components/BalanceTracker.tsx
@@ -12,7 +12,7 @@ import {
 } from '@mui/material';
 import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { CurrencyField } from 'src/components/common/forms';
+import { CurrencyField } from 'src/components/forms';
 import { IOperation, ITransaction } from 'src/models/finances';
 import PriceChangeIcon from '@mui/icons-material/PriceChange';
 import { useDialogFullScreen } from 'src/hooks';
