+++ src/pages/client/_finances/TradingJournal/components/OperationsHistory.tsx
