+++ src/pages/client/_finances/TradingJournal/components/OperationsHistoryChart.tsx
