+++ src/pages/client/_finances/TradingJournal/components/SummaryPanel.tsx
