+++ src/pages/client/_finances/TradingJournal/components/Timeline.tsx
