+++ src/pages/client/_finances/TradingJournal/components/index.ts
