+++ src/pages/client/_finances/TradingJournal/index.ts
@@ -1,0 +1,1 @@
+export { default as TradingJournalPage } from './Page';
