+++ src/pages/client/_finances/index.ts
@@ -1,0 +1,4 @@
+export { PersonalFinancesPage } from './PersonalFinances';
+export { TradingJournalPage } from './TradingJournal';
+export { PortfolioPage } from './Portfolio';
+export { CompoundCalculatorPage } from './CompoundCalculator';
