--- src/pages/client/finances/index.ts
@@ -1,4 +1,0 @@
-export { default as PersonalFinancesPage } from './PersonalFinancesPage';
-export { default as PortfolioPage } from './Portfolio';
-export { default as TradingJournalPage } from './TradingJournalPage';
-export { CompoundCalculatorPage } from './CompoundCalculator';
